requirements.txt
torch==1.8.0
torchvision==0.9.0
scikit-learn
numpy
Pillow
Django==3.2.7
